fix circular dependency
https://coveord.atlassian.net/browse/KIT-316
